test(backend): lock automation and project git branch name builders
Extract pure helpers so Vitest can cover per-run automation branches and v2+ project suffixes without Convex auth imports.
